+++ addOns/sequence/src/main/java/org/zaproxy/zap/extension/sequence/ExtensionSequence.java
@@ -21,8 +21,8 @@ package org.zaproxy.zap.extension.sequence;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 import javax.swing.ImageIcon;
+import org.apache.commons.configuration.ConfigurationException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.control.Control;
@@ -36,12 +36,12 @@ import org.parosproxy.paros.extension.ViewDelegate;
 import org.parosproxy.paros.network.HttpMessage;
 import org.zaproxy.addon.network.ExtensionNetwork;
 import org.zaproxy.zap.extension.ascan.ExtensionActiveScan;
+import org.zaproxy.zap.extension.ascan.ScanPolicy;
 import org.zaproxy.zap.extension.script.ExtensionScript;
-import org.zaproxy.zap.extension.script.ScriptCollection;
 import org.zaproxy.zap.extension.script.ScriptType;
 import org.zaproxy.zap.extension.script.ScriptWrapper;
-import org.zaproxy.zap.extension.script.SequenceScript;
 import org.zaproxy.zap.extension.zest.ExtensionZest;
+import org.zaproxy.zap.extension.zest.ZestScriptWrapper;
 
 public class ExtensionSequence extends ExtensionAdaptor implements ScannerHook {
 
@@ -57,6 +57,8 @@ public class ExtensionSequence extends ExtensionAdaptor implements ScannerHook {
     private SequenceAscanPanel sequencePanel;
 
     private ScriptType scriptType;
+
+    private boolean scanning = false;
 
     public ExtensionSequence() {
         super("ExtensionSequence");
@@ -116,10 +118,43 @@ public class ExtensionSequence extends ExtensionAdaptor implements ScannerHook {
         getExtScript().removeScriptType(scriptType);
     }
 
+    public ScanPolicy getDefaultScanPolicy() throws ConfigurationException {
+        // FIXME: this should be read from the options
+        return getExtActiveScan().getPolicyManager().getPolicy("Sequence");
+    }
+
     @Override
     public void scannerComplete() {
-        // Reset the sequence extension
-        this.directScripts = null;
+        if (!scanning && sequencePanel != null) {
+            // Active scan all of the selected sequences
+            // Need this bool otherwise we will infinitely recurse!
+            scanning = true;
+            List<ScriptWrapper> scripts = sequencePanel.getPanel(false).getSelectedIncludeScripts();
+
+            List<Object> contextSpecificObjects = new ArrayList<>();
+            try {
+                contextSpecificObjects.add(getDefaultScanPolicy());
+            } catch (ConfigurationException e4) {
+                // Ignore
+            }
+            scripts.forEach(
+                    s -> {
+                        if (s instanceof ZestScriptWrapper) {
+                            StdActiveScanRunner zzr =
+                                    new StdActiveScanRunner(
+                                            (ZestScriptWrapper) s,
+                                            null,
+                                            null,
+                                            contextSpecificObjects);
+                            try {
+                                zzr.run(null, null);
+                            } catch (Exception e1) {
+                                LOGGER.error(e1.getMessage(), e1);
+                            }
+                        }
+                    });
+            scanning = false;
+        }
     }
 
     @Override
@@ -148,74 +183,14 @@ public class ExtensionSequence extends ExtensionAdaptor implements ScannerHook {
     }
 
     @Override
-    public void beforeScan(HttpMessage msg, AbstractPlugin plugin, Scanner scanner) {
-        // If the HttpMessage has a HistoryReference with an ID that is also in the HashMap of the
-        // Scanner,
-        // then the message has a specific Sequence script to scan.
-        SequenceScript seqScr = getIncludedSequenceScript(msg, scanner);
-
-        // If any script was found, send all the requests prior to the message to be scanned.
-        if (seqScr != null) {
-            HttpMessage newMsg = seqScr.runSequenceBefore(msg, plugin);
-            updateMessage(msg, newMsg);
-        }
-    }
+    public void beforeScan(HttpMessage msg, AbstractPlugin plugin, Scanner scanner) {}
 
     @Override
-    public void afterScan(HttpMessage msg, AbstractPlugin plugin, Scanner scanner) {
-        // If the HttpMessage has a HistoryReference with an ID that is also in the HashMap of the
-        // Scanner,
-        // then the message has a specific Sequence script to scan.
-        SequenceScript seqScr = getIncludedSequenceScript(msg, scanner);
-
-        // If any script was found, send all the requests after the message that was scanned.
-        if (seqScr != null) {
-            seqScr.runSequenceAfter(msg, plugin);
-        }
-    }
-
-    private SequenceScript getIncludedSequenceScript(HttpMessage msg, Scanner scanner) {
-        List<ScriptWrapper> sequences = directScripts;
-        if (sequences == null) {
-            Set<ScriptCollection> scs = scanner.getScriptCollections();
-            if (scs != null) {
-                for (ScriptCollection sc : scs) {
-                    if (sc.getType().getName().equals(TYPE_SEQUENCE)) {
-                        sequences = sc.getScripts();
-                        break;
-                    }
-                }
-            }
-        }
-        if (sequences != null) {
-            for (ScriptWrapper wrapper : sequences) {
-                try {
-                    SequenceScript seqScr =
-                            getExtScript().getInterface(wrapper, SequenceScript.class);
-                    if (seqScr != null) {
-                        if (seqScr.isPartOfSequence(msg)) {
-                            return seqScr;
-                        }
-                    }
-                } catch (Exception e) {
-                    LOGGER.debug(
-                            "Exception occurred, while trying to fetch Included Sequence Script: {}",
-                            e.getMessage());
-                }
-            }
-        }
-        return null;
-    }
+    public void afterScan(HttpMessage msg, AbstractPlugin plugin, Scanner scanner) {}
 
     public void setDirectScanScript(ScriptWrapper script) {
         directScripts = new ArrayList<>();
         directScripts.add(script);
-    }
-
-    private void updateMessage(HttpMessage msg, HttpMessage newMsg) {
-        msg.setRequestHeader(newMsg.getRequestHeader());
-        msg.setRequestBody(newMsg.getRequestBody());
-        msg.setCookies(new ArrayList<>());
     }
 
     private ExtensionScript getExtScript() {